fix(den-db): address Stage 1 review findings (mig-0, arch-0/1/2, ts-0/1)
Zero-findings pass after a 5-lens review of the Stage 1 schema commit.

- mig-0: run the idempotent FULLTEXT ensure on ALL apply paths, not just
  bootstrap. `db:migrate` and `db:push` now chain `ensure-fulltext-indexes.ts`,
  so a `db:migrate`-only run on an already-baselined DB (the prod path in
  den-db-migrate.yml, and the local db:push dev path) still creates the index.
- arch-0: expose the proof harness as `db:verify-memory` (was only runnable via
  a raw node invocation, weakening its CI-gate claim).
- arch-1: single `ensureFulltextIndexes()` seam in fulltext.ts called by both
  apply paths, so they cannot drift; bootstrap log is now accurate.
- arch-2: extract createExecutor/Executor into scripts/db-executor.ts, shared by
  bootstrap.ts and the ensure runner (removes memory-scoped triplication).
- ts-0/int-0: roll back before best-effort cleanup in verify-memory-schema.ts so
  a mid-transaction failure cannot leak marker rows.
- ts-1: make isRecord file-private (dead export surface).

Re-proven against MySQL 8.4: bootstrap, db:push, and the db:migrate chain each
create the FULLTEXT index; second ensure is a no-op; db:verify-memory passes.
